+++ lang/src/org/partiql/lang/prettyprint/ASTPrettyPrinter.kt
@@ -394,33 +394,8 @@ class ASTPrettyPrinter {
                 attrOfParent = attrOfParent,
                 children = toRecursionTreeList(node.operands)
             )
-            is PartiqlAst.Expr.Union -> RecursionTree(
-                astType = "Union",
-                attrOfParent = attrOfParent,
-                children = toRecursionTreeList(node.operands)
-            )
-            is PartiqlAst.Expr.Intersect -> RecursionTree(
-                astType = "Intersect",
-                attrOfParent = attrOfParent,
-                children = toRecursionTreeList(node.operands)
-            )
-            is PartiqlAst.Expr.Except -> RecursionTree(
-                astType = "Except",
-                attrOfParent = attrOfParent,
-                children = toRecursionTreeList(node.operands)
-            )
-            is PartiqlAst.Expr.OuterUnion -> RecursionTree(
-                astType = "OuterUnion",
-                attrOfParent = attrOfParent,
-                children = toRecursionTreeList(node.operands)
-            )
-            is PartiqlAst.Expr.OuterIntersect -> RecursionTree(
-                astType = "OuterIntersect",
-                attrOfParent = attrOfParent,
-                children = toRecursionTreeList(node.operands)
-            )
-            is PartiqlAst.Expr.OuterExcept -> RecursionTree(
-                astType = "OuterExcept",
+            is PartiqlAst.Expr.BagOp -> RecursionTree(
+                astType = node.op.javaClass.simpleName.capitalize(),
                 attrOfParent = attrOfParent,
                 children = toRecursionTreeList(node.operands)
             )